chore: 인터페이스 types 파일로 분리
