+++ src/types/project.ts
@@ -1,0 +1,10 @@
+interface Project {
+  id: number;
+  name: string;
+  description: string;
+  type: string;
+  generation: number;
+  thumbnailUrl: string;
+}
+
+export type { Project };
